refactor: remove internal CLI dispatch barrel
