+++ packages/cli/src/commands/chat.ts
@@ -4,7 +4,7 @@ import { assertExplicitModelKnown } from '../runtime/runModel';
 import { notifyCliUpdate } from '../runtime/updateChecker';
 
 import { contextFromArgs } from './_helpers/context';
-import { withUsageSections } from './_helpers/dispatch';
+import { withUsageSections } from './_helpers/dispatch/usage';
 import { setExitCode } from './_helpers/exitCode';
 import {
   INTERACTIVE_AGENT_GLOBAL_ARGS,
