+++ packages/cli/src/commands/setup.ts
@@ -9,7 +9,7 @@ import {
 } from '../runtime/terminalRequirements';
 
 import { contextFromArgs } from './_helpers/context';
-import { withUsageSections } from './_helpers/dispatch';
+import { withUsageSections } from './_helpers/dispatch/usage';
 import { setExitCode } from './_helpers/exitCode';
 import {
   INTERACTIVE_GLOBAL_ARGS,
